refactor: filters popover mounted outside data table
